pyproject.toml
[project]
name = "wts"
version = "0.1.0"
description = "Search phrases across YouTube channel transcripts"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Gnanadeep" }
]

dependencies = [
    "requests",
    "python-dotenv",
    "yt-dlp",
    "youtube-transcript-api",
]
